requirements.txt
altair==5.3.0
attrs==23.2.0
blinker==1.9.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
filelock==3.19.1
Flask==3.1.3
flask-cors==6.0.2
fsspec==2025.9.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
networkx==3.5
numpy==1.26.4
packaging==26.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
psycopg2-binary==2.9.11
pyarrow==16.1.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
referencing==0.35.1
requests==2.31.0
rich==13.7.1
setuptools==82.0.1
six==1.17.0
streamlit==1.34.0
torch==2.8.0
torchvision==0.23.0
Werkzeug==3.1.8
wheel==0.46.3
